% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dati_meteo.R
\docType{data}
\name{dati_meteo}
\alias{dati_meteo}
\title{Data of nominal brown shrimp prices by size}
\format{
A tibble with 243 rows and 10 variables:
\describe{
\item{station_code}{chr codice stazione}
\item{station_eu_code}{chr codice europeo della stazione}
\item{date}{date data}
\item{t2m}{dbl temperatura media a 2m}
\item{tmin2m}{dbl temperatura minima a 2m}
\item{tmax2m}{dbl temperatura massima a 2m}
\item{tp}{dbl precipitazione totale}
\item{ptp}{dbl precipitazione totale del giorno precedente}
\item{rh}{dbl umidità relativa}
\item{u10m}{dbl componente U per la direzione del vento}
\item{v10m}{dbl componente V per la direzione del vento}
\item{sp}{dbl pressione}
\item{nirradiance}{dbl net radiance}
\item{pbl00}{dbl Planet Boundary Layer ore 00}
\item{pbl12}{dbl Planet Boundary Layer ore 12	}
\item{pblmin}{dbl Planet Boundary Layer daily minimum	}
\item{pblmax}{dbl Planet Boundary Layer daily maximum	}
\item{coordx}{dbl latitudine EPSG:32632  (WGS84 UTM32)}
\item{coordy}{dbl longitudine EPSG:32632  (WGS84 UTM32)}
\item{altitude}{chr}
\item{altitudedem}{chr}
}
}
\source{
\url{https://www.pnas.org/content/114/7/1512/tab-figures-data}
}
\usage{
dati_meteo
}
\description{
Shrimp are classified by size, 0-15 shrimp per pound, 15-20 shrimp per pound, etc. A smaller number per pound indicates larger shrimp. Nominal prices are total monthly value of brown shrimp andings within size class divided by total monthly landings within the size class.
}
\keyword{datasets}
