% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datiMeteo-package.R
\docType{package}
\name{datiMeteo-package}
\alias{datiMeteo}
\alias{datiMeteo-package}
\title{datiMeteo: I dati per gli inquinanti atmosferici del progetto Pulvirus}
\description{
Il pacchetto contiene le serie storiche degli inquinanti dal 2016 al 2020.
    Contiene inoltre l'anagrafica delle stazioni di monitoraggio ritenute valide secondo i criteri stabiliti. 
    Una overview del contenuto si ottiene con data(package = "datiInquinanti").
}
\author{
\strong{Maintainer}: Raffaele Morelli \email{raffaele.morelli@isprambiente.it} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
