% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stazioniMeteo.R
\docType{data}
\name{stazioniMeteo}
\alias{stazioniMeteo}
\title{Anagrafica delle stazioni ricavata dal dataset dei dati meteorologici}
\format{
A tibble:
\describe{
\item{station_eu_code}{char codice europeo della stazione}
\item{coordx}{dbl latitudine EPSG:32632 (WGS84 UTM32)}
\item{coordy}{dbl longitudine EPSG:32632 (WGS84 UTM32)}
\item{altitude}{dbl altitudine}
\item{altitudedem}{dbl altitudine ricavata dal DEM}
}
}
\usage{
stazioniMeteo
}
\description{
Anagrafica delle stazioni ricavata dal dataset dei dati meteorologici
}
\keyword{datasets}
